+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/impl/SchemaCleanUpUtils.java
@@ -19,14 +19,21 @@ package com.github.victools.jsonschema.generator.impl;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.fasterxml.jackson.databind.node.TextNode;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfig;
 import com.github.victools.jsonschema.generator.SchemaKeyword;
 import com.github.victools.jsonschema.generator.SchemaVersion;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.EnumSet;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.BinaryOperator;
 import java.util.function.Consumer;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
@@ -48,23 +55,23 @@ public class SchemaCleanUpUtils {
     }
 
     /**
-     * Remove and merge {@link SchemaKeyword#TAG_ALLOF} parts when there are no conflicts between the sub-schemas.
-     * <br>
-     * This makes for more readable schemas being generated but has the side-effect that manually added {@link SchemaKeyword#TAG_ALLOF} (e.g. from a
-     * custom definition or attribute overrides) may be removed as well if it isn't strictly speaking necessary.
+     * Remove and merge {@link SchemaKeyword#TAG_ALLOF} parts when there are no conflicts between the sub-schemas. This makes for more readable
+     * schemas being generated but has the side-effect that any manually added {@link SchemaKeyword#TAG_ALLOF} (e.g. through a custom definition of
+     * attribute overrides) may be removed as well if it isn't strictly speaking necessary.
      *
      * @param jsonSchemas generated schemas that may contain unnecessary {@link SchemaKeyword#TAG_ALLOF} nodes
      */
     public void reduceAllOfNodes(List<ObjectNode> jsonSchemas) {
         String allOfTagName = this.config.getKeyword(SchemaKeyword.TAG_ALLOF);
-        this.finaliseSchemaParts(jsonSchemas, nodeToCheck -> this.mergeAllOfPartsIfPossible(nodeToCheck, allOfTagName));
+        Map<String, SchemaKeyword> reverseKeywordMap = EnumSet.allOf(SchemaKeyword.class).stream()
+                .collect(Collectors.toMap(this.config::getKeyword, keyword -> keyword));
+        this.finaliseSchemaParts(jsonSchemas, nodeToCheck -> this.mergeAllOfPartsIfPossible(nodeToCheck, allOfTagName, reverseKeywordMap));
     }
 
     /**
-     * Reduce nested {@link SchemaKeyword#TAG_ANYOF} parts when one contains an entry with only another {@link SchemaKeyword#TAG_ANYOF} inside.
-     * <br>
-     * This makes for more readable schemas being generated but has the side-effect that manually added {@link SchemaKeyword#TAG_ANYOF} (e.g. from a
-     * custom definition or attribute overrides) may be removed as well if it isn't strictly speaking necessary.
+     * Reduce nested {@link SchemaKeyword#TAG_ANYOF} parts when one contains an entry with only another {@link SchemaKeyword#TAG_ANYOF} inside. This
+     * makes for more readable schemas being generated but has the side-effect that any manually added {@link SchemaKeyword#TAG_ANYOF} (e.g. through a
+     * custom definition of attribute overrides) may be removed as well if it isn't strictly speaking necessary.
      *
      * @param jsonSchemas generated schemas that may contain unnecessary nested {@link SchemaKeyword#TAG_ANYOF} nodes
      */
@@ -162,8 +169,9 @@ public class SchemaCleanUpUtils {
      *
      * @param schemaNode single node representing a sub-schema to consolidate contained {@link SchemaKeyword#TAG_ALLOF} for (if present)
      * @param allOfTagName name of the {@link SchemaKeyword#TAG_ALLOF} in the designated JSON Schema version
+     * @param reverseKeywordMap mapping from actual tag name in generated schema to underlying {@link SchemaKeyword}
      */
-    private void mergeAllOfPartsIfPossible(JsonNode schemaNode, String allOfTagName) {
+    private void mergeAllOfPartsIfPossible(JsonNode schemaNode, String allOfTagName, Map<String, SchemaKeyword> reverseKeywordMap) {
         if (!(schemaNode instanceof ObjectNode)) {
             return;
         }
@@ -171,7 +179,7 @@ public class SchemaCleanUpUtils {
         if (!(allOfTag instanceof ArrayNode)) {
             return;
         }
-        allOfTag.forEach(part -> this.mergeAllOfPartsIfPossible(part, allOfTagName));
+        allOfTag.forEach(part -> this.mergeAllOfPartsIfPossible(part, allOfTagName, reverseKeywordMap));
 
         List<JsonNode> allOfElements = new ArrayList<>();
         allOfTag.forEach(allOfElements::add);
@@ -182,45 +190,158 @@ public class SchemaCleanUpUtils {
                 .filter(part -> part instanceof ObjectNode)
                 .map(part -> (ObjectNode) part)
                 .collect(Collectors.toList());
+
         // collect all defined attributes from the separate parts and check whether there are incompatible differences
         Map<String, List<JsonNode>> fieldsFromAllParts = Stream.concat(Stream.of(schemaNode), parts.stream())
                 .flatMap(part -> StreamSupport.stream(((Iterable<Map.Entry<String, JsonNode>>) () -> part.fields()).spliterator(), false))
-                .collect(Collectors.groupingBy(Map.Entry::getKey, Collectors.mapping(Map.Entry::getValue, Collectors.toList())));
+                .collect(Collectors.groupingBy(Map.Entry::getKey, LinkedHashMap::new, Collectors.mapping(Map.Entry::getValue, Collectors.toList())));
         if ((this.config.getSchemaVersion() == SchemaVersion.DRAFT_6 || this.config.getSchemaVersion() == SchemaVersion.DRAFT_7)
                 && fieldsFromAllParts.containsKey(this.config.getKeyword(SchemaKeyword.TAG_REF))
                 && (schemaNode.size() > 1 || parts.size() > 1)) {
             // in Draft 7, any other attributes besides the $ref keyword were ignored
             return;
         }
-        String ifTagName = this.config.getKeyword(SchemaKeyword.TAG_IF);
-        for (Map.Entry<String, List<JsonNode>> fieldEntries : fieldsFromAllParts.entrySet()) {
-            if (ifTagName.equals(fieldEntries.getKey())) {
+        Map<String, List<JsonNode>> unsupportedTagValues = fieldsFromAllParts.entrySet().stream()
+                .filter(entry -> !reverseKeywordMap.containsKey(entry.getKey()))
+                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, throwingMerger(), LinkedHashMap::new));
+        if (unsupportedTagValues.entrySet().stream().anyMatch(entry -> entry.getValue().size() > 1)) {
+            // unsupported tag with more than one occurrence: be conservative and don't merge
+            return;
+        }
+        Map<SchemaKeyword, List<JsonNode>> supportedTagValues = fieldsFromAllParts.entrySet().stream()
+                .filter(entry -> !unsupportedTagValues.containsKey(entry.getKey()))
+                .collect(Collectors.toMap(entry -> reverseKeywordMap.get(entry.getKey()), Map.Entry::getValue, throwingMerger(), LinkedHashMap::new));
+        if (supportedTagValues.containsKey(SchemaKeyword.TAG_IF)) {
                 // "if"/"then"/"else" tags should remain isolated in their sub-schemas
-                return;
-            }
-            if (fieldEntries.getValue().size() == 1) {
-                // no conflicts, no further checks
-                continue;
-            }
-            int offset;
-            if (!allOfTagName.equals(fieldEntries.getKey())) {
-                offset = 0;
-            } else if (fieldEntries.getValue().size() == 2) {
+            return;
+        }
+        Map<SchemaKeyword, Supplier<JsonNode>> supportedTagValueSuppliers = new LinkedHashMap<>();
+        for (Map.Entry<SchemaKeyword, List<JsonNode>> fieldEntries : supportedTagValues.entrySet()) {
+            SchemaKeyword keyword = fieldEntries.getKey();
+            List<JsonNode> valuesToMerge = fieldEntries.getValue();
+            if (keyword == SchemaKeyword.TAG_ALLOF) {
                 // we can ignore the "allOf" tag in the target node (the one we are trying to remove here)
-                continue;
-            } else {
-                offset = 1;
+                valuesToMerge = valuesToMerge.subList(1, valuesToMerge.size());
+                if (valuesToMerge.isEmpty()) {
+                    // no other "allOf" part left to merge
+                    continue;
+                }
             }
-            if (!fieldEntries.getValue().stream().skip(offset + 1).allMatch(fieldEntries.getValue().get(offset)::equals)) {
-                // different values for the same tag: be conservative for now and not merge anything
-                // later, we may want to decide based on the tag how to merge them (e.g. take the highest "minLength" and the lowest "maximum")
+            Supplier<JsonNode> mergeResultSupplier = this.getAllOfMergeFunctionFor(keyword, valuesToMerge);
+            if (mergeResultSupplier == null) {
+                // unable to merge given values for the specified keyword; abort merge
                 return;
             }
+            supportedTagValueSuppliers.put(keyword, mergeResultSupplier);
         }
         // all attributes are either distinct or have equal values in all occurrences
         ObjectNode schemaObjectNode = (ObjectNode) schemaNode;
         schemaObjectNode.remove(allOfTagName);
-        parts.forEach(schemaObjectNode::setAll);
+        // for the supported tags, there may be multiple occurrences that requiring individual merging
+        supportedTagValueSuppliers.forEach((keyword, valueSupplier) -> schemaObjectNode.set(this.config.getKeyword(keyword), valueSupplier.get()));
+        // for the unsupported tags, there is only a single occurrence each
+        unsupportedTagValues.forEach((tagName, valueList) -> schemaObjectNode.set(tagName, valueList.get(0)));
+    }
+
+    private Supplier<JsonNode> getAllOfMergeFunctionFor(SchemaKeyword keyword, List<JsonNode> valuesToMerge) {
+        if (valuesToMerge.size() == 1) {
+            // no conflicts, no further checks
+            return () -> valuesToMerge.get(0);
+        }
+        switch (keyword) {
+        case TAG_ALLOF:
+        case TAG_REQUIRED:
+            return this.mergeArrays(valuesToMerge);
+        case TAG_PROPERTIES:
+            return this.mergeObjectProperties(valuesToMerge);
+        case TAG_TYPE:
+            return this.returnOverlapOfStringsOrStringArrays(valuesToMerge);
+        default:
+            return this.returnOneIfAllEqual(valuesToMerge);
+        }
+    }
+
+    private Supplier<JsonNode> mergeArrays(List<JsonNode> arrayNodesToMerge) {
+        if (!arrayNodesToMerge.stream().allMatch(JsonNode::isArray)) {
+            // at least one value is not an array as expected, abort merge
+            return null;
+        }
+        return () -> {
+            ArrayNode mergedArrayNode = this.config.createArrayNode();
+            arrayNodesToMerge.forEach(node -> node.forEach(mergedArrayNode::add));
+            return mergedArrayNode;
+        };
+    }
+
+    private Supplier<JsonNode> mergeObjectProperties(List<JsonNode> objectNodesToMerge) {
+        if (!objectNodesToMerge.stream().allMatch(JsonNode::isObject)) {
+            // at least one value is not an object as expected, abort merge
+            return null;
+        }
+        ObjectNode mergedObjectNode = this.config.createObjectNode();
+        for (JsonNode singleObjectNode : objectNodesToMerge) {
+            Iterator<Map.Entry<String, JsonNode>> it = singleObjectNode.fields();
+            while (it.hasNext()) {
+                Map.Entry<String, JsonNode> singleField = it.next();
+                if (!mergedObjectNode.has(singleField.getKey())) {
+                    mergedObjectNode.set(singleField.getKey(), singleField.getValue());
+                } else if (mergedObjectNode.get(singleField.getKey()).equals(singleField.getValue())) {
+                    // preserve equal existing node
+                    continue;
+                } else {
+                    // cannot consolidate two occurrences of the same property; abort merge (in the future: may want to be smarter here)
+                    return null;
+                }
+            }
+        }
+        return () -> mergedObjectNode;
+    }
+
+    private Supplier<JsonNode> returnOverlapOfStringsOrStringArrays(List<JsonNode> nodes) {
+        List<String> encounteredValues = this.getStringValuesFromStringOrStringArray(nodes.get(0));
+        if (encounteredValues == null) {
+            return null;
+        }
+        for (JsonNode nextNode : nodes.subList(1, nodes.size())) {
+            List<String> nextValues = this.getStringValuesFromStringOrStringArray(nextNode);
+            if (nextValues == null) {
+                // invalid node; abort merge
+                return null;
+            }
+            encounteredValues.retainAll(nextValues);
+            if (encounteredValues.isEmpty()) {
+                // invalid merge result: no valid value remained; abort merge
+                return null;
+            }
+        }
+        if (encounteredValues.size() == 1) {
+            return () -> new TextNode(encounteredValues.get(0));
+        }
+        return () -> this.config.createArrayNode()
+                .addAll(encounteredValues.stream().map(TextNode::new).collect(Collectors.toList()));
+    }
+
+    private List<String> getStringValuesFromStringOrStringArray(JsonNode node) {
+        if (node.isArray()) {
+            List<String> result = new ArrayList<>();
+            node.forEach(arrayItem -> result.add(arrayItem.asText(null)));
+            if (result.contains(null)) {
+                return null;
+            }
+            return result;
+        }
+        if (node.isTextual()) {
+            return Collections.singletonList(node.asText());
+        }
+        // neither array nor text node; abort merge
+        return null;
+    }
+
+    private Supplier<JsonNode> returnOneIfAllEqual(List<JsonNode> nodes) {
+        if (nodes.subList(1, nodes.size()).stream().allMatch(nodes.get(0)::equals)) {
+            return () -> nodes.get(0);
+        }
+        return null;
     }
 
     /**
@@ -294,5 +415,17 @@ public class SchemaCleanUpUtils {
                 .replaceAll(",", ".")
                 // removing white-spaces and any other remaining invalid characters
                 .replaceAll("[^a-zA-Z0-9\\.\\-_]+", "");
+    }
+
+    /**
+     * Helper function, that returns a BinaryOperator for use in Collectors.toMap() that assumes that there are no duplicate keys.
+     *
+     * @param <T> value type
+     * @return merge function that throws an IllegalStateException when invoked
+     */
+    private static <T> BinaryOperator<T> throwingMerger() {
+        return (u,v) -> {
+            throw new IllegalStateException("Duplicate key " + u);
+        };
     }
 }